api_server/requirements.txt
flask
flask-socketio
flask-cors
eventlet
torch
torchvision
opencv-python
numpy
Pillow
albumentations
mediapipe